DRY out commonality between the Noble and WASM versions.
